Kong dev
06 june 

update block_dev after updated,
idead - config map and requirement by p'mott , change dashboard, change the config provice and camp.
kong - fix the row can't insert all row, change the rule how to insert into log_act_detail all
list - Actvities_header
- land_code
- camp
- act_header_type
- act_header_type_detail
- resource_type
- resource_item
- units
- check duplicate a log_act_detail

+++ managementDataSystem_forCalculate_1.3_05192026_postgres.sql
@@ -329,7 +329,7 @@ CREATE TABLE "log_activities_detail" (
   "resource_used_type_id" integer,
   "unit_prefix_id" integer,
   "unit_id" integer,
-  "log_act_detail_quatity" int,
+  "log_act_detail_quatity" float,
   "log_act_detail_volumePerUnit" float,
   "log_act_detail_volumeAll" float,
   "log_act_detail_areawork" float,
